Concat engine into one file.
