--- lib/cls.js
@@ -1,5 +1,0 @@
-export default function(ctx) {
-  ctx.globalCompositeOperation = 'source-over';
-  ctx.setTransform(1, 0, 0, 1, 0, 0);
-  ctx.clearRect(0, 0, ctx.canvas.clientWidth, ctx.canvas.clientHeight);
-}
