+++ lib/engine.js
@@ -1,3 +1,5 @@
+import mapinfo from "../assets/map";
+
 const fontResource = document.getElementById('font');
 
 const canvas = document.createElement('canvas');
@@ -245,7 +247,7 @@ function getSize() {
   };
 }
 
-function posFromIndex(index) {
+function mapPosFromIndex(index) {
   return [
     (index % spriteSheetColumns) * spriteSheetTileWidth,
     Math.floor(index / spriteSheetColumns) * spriteSheetTileHeight,
@@ -277,7 +279,7 @@ function draw(ctx, name, x, y) {
 }
 
 function drawTile(ctx, id, x, y) {
-  const [tileX, tileY] = posFromIndex(id);
+  const [tileX, tileY] = mapPosFromIndex(id);
   ctx.drawImage(
     spriteSheetImg,
     tileX,
@@ -291,10 +293,240 @@ function drawTile(ctx, id, x, y) {
   );
 }
 
+
+// PROCESS
+
+const processAllSet = new Set();
+
+function process(update = () => {}, render = () => {}) {
+  let accumulatedTime = 0;
+  let lastTime = 0;
+  let deltaTime = 1 / 60;
+
+  function tick(time) {
+    accumulatedTime += (time - lastTime) / 1000;
+
+    if (accumulatedTime > 1) {
+      accumulatedTime = 1;
+    }
+
+    while (accumulatedTime > deltaTime) {
+      update(deltaTime);
+      accumulatedTime -= deltaTime;
+    }
+
+    render(accumulatedTime / deltaTime);
+    lastTime = time;
+  }
+
+  processAllSet.add(tick);
+}
+
+function processAll(time) {
+  requestAnimationFrame(processAll);
+  processAllSet.forEach(process => {
+    process(time);
+  });
+}
+
+function start() {
+  requestAnimationFrame(processAll);
+}
+
+// COOLDOWN
+
+const cooldowns = new Set();
+let time = 0;
+
+function cooldown() {
+  const map = new Map();
+  cooldowns.add(map);
+
+  function has(name) {
+    if (!map.has(name)) {
+      return false;
+    }
+
+    const { value, created } = map.get(name);
+    return time - created < value;
+  }
+
+  function set(name, value = 1) {
+    map.set(name, {
+      value: value,
+      created: time,
+    });
+  }
+
+  function hasSet(name, value) {
+    if (has(name)) {
+      return true;
+    } else {
+      set(name, value);
+      return false;
+    }
+  }
+
+  return {
+    has,
+    set,
+    hasSet,
+  };
+}
+
+process(dt => {
+  time += dt;
+});
+
+// CLS
+
+function cls(ctx) {
+  ctx.globalCompositeOperation = 'source-over';
+  ctx.setTransform(1, 0, 0, 1, 0, 0);
+  ctx.clearRect(0, 0, ctx.canvas.clientWidth, ctx.canvas.clientHeight);
+}
+
+// CAMERA
+
+const cameraViewport = {
+  width: 128,
+  height: 128,
+  originx: 64,
+  originy: 64,
+};
+
+function camera(ctx, x = 0, y = 0, rotation = 0, scale = 1) {
+  const rads = toRadians(rotation);
+  ctx.resetTransform();
+  ctx.translate(cameraViewport.originx, cameraViewport.originy);
+  ctx.rotate(rads);
+  ctx.scale(scale, scale);
+  ctx.translate(-cameraViewport.originx, -cameraViewport.originy);
+  ctx.translate(-x, -y);
+}
+
+// MAP
+
+
+const mapTileNormals = [
+  { x: 0, y: -1 },
+  { x: 1, y: 0 },
+  { x: 0, y: 1 },
+  { x: -1, y: 0 },
+];
+
+const mapSpriteSize = getSize();
+
+// For some reason tileset now has width of 32 and height of 16.
+// So for now instead of relying on size values from map, it's fixed here.
+const mapSize = 16;
+
+// This is fever dream based collision detection.
+// Idea: Checking which edge normal is aligned best
+// Sources:
+// - Reflection: https://math.stackexchange.com/questions/13261/how-to-get-a-reflection-vector
+function getCollider(x, y, lastX, lastY, nextVelocity) {
+  // Collider map
+  const cmap = getCollisionMap();
+
+  const tx = Math.floor((x / mapSpriteSize.width) * mapSize);
+  const ty = Math.floor((y / mapSpriteSize.height) * mapSize);
+  const worldTileSize = mapSpriteSize.tileWidth;
+
+  const index = tx + ty * cmap.width;
+  const value = cmap.data[index];
+
+  if (!value) {
+    return null;
+  }
+
+  // Tile Word position X
+  const twx = (tx / mapSize) * mapSpriteSize.width;
+  const twy = (ty / mapSize) * mapSpriteSize.height;
+
+  const center = worldTileSize / 2;
+
+  // Offsetting vertices to the center of each edge for better dot resolution where looking for most aligned edge normal
+  const vertices = [
+    { x: twx + center, y: twy },
+    { x: twx + worldTileSize, y: twy + center },
+    {
+      x: twx + worldTileSize - center,
+      y: twy + worldTileSize,
+    },
+    { x: twx, y: twy + worldTileSize - center },
+  ];
+
+  var closestDot = -Number.MAX_VALUE;
+  var closestIndex = 0;
+
+  const lastPos = { x: lastX, y: lastY };
+
+  // Think about sub-stepping!
+  for (let i = 0; i < 4; i++) {
+    const dir = vecSub(lastPos, vertices[i]);
+    const dot = vecDot(dir, mapTileNormals[i]);
+
+    // Take into account if specific edge is facing another tile. If so, block it as possible selection.
+    // LAAAAAMEEE! We should validate against object being inside square. But whatever... works 99% of the time ;)
+    if (dot > closestDot) {
+      closestDot = dot;
+      closestIndex = i;
+    }
+  }
+
+  const closestNormal = mapTileNormals[closestIndex];
+
+  return {
+    code: value,
+    tx: tx,
+    ty: ty,
+    x: twx + worldTileSize / 2,
+    y: twy + worldTileSize / 2,
+    velocity: vecReflect(nextVelocity, closestNormal),
+  };
+}
+
+function getCollisionMap() {
+  return mapinfo.layers.find(layer => {
+    return layer.name === 'collision';
+  });
+}
+
+function map(ctx) {
+  mapinfo.layers.forEach(layer => {
+    const { width, data, name } = layer;
+
+    if (name === 'collision') {
+      if (!$globalConfig.isDebugDraw) {
+        return;
+      }
+      ctx.globalAlpha = 0.3;
+    } else {
+      ctx.globalAlpha = 1;
+    }
+
+    data.forEach((tile, index) => {
+      if (tile === 0) return;
+      drawTile(
+        ctx,
+        tile - 1,
+        index % width,
+        Math.floor(index / width)
+      );
+    });
+  });
+}
+
 export default {
   print,
   printOutline,
   init, isUp, isDown, isDownOnce, clear,
   drawGrid, drawRect, drawCircle, drawLine, lerp, vecDot, vecSub, vecLength, vecFromAngle, vecMultScalar, vecReflect, toRadians,
-  define, draw, drawTile, getSize
+  define, draw, drawTile, getSize,
+  start, process, processAll,
+  cooldown,
+  cls,
+  camera,
+  map, getCollisionMap, getCollider
 };
