+++ src/fx.js
@@ -1,5 +1,3 @@
-import process from '/lib/process';
-
 const particles = new Set();
 
 for (let i = 0; i < 200; i++) {
